Issue#14: Moved the handling of the JFrame for the debug window into the Main so that it is a standard feature.
